video_annotation_platform/requirements.txt
anthropic>=0.40.0
opencv-python>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
ultralytics>=8.0.0
transformers>=4.37.0
torch>=2.1.0
tqdm>=4.66.0
pyyaml>=6.0
